Update the hypothesis for the aggregate
Instead of figuring the step which frankly can happen with a single SQL
expression (value - lag(value, 1) over w) without even bothering with a CTE try
to set up a rather convoluted scenario that expects three column outputs.

Commulative sum, step (defined as current value - previous value), and running
max step (defined as the overall maximum step until now, unless now is the
overall maximum)

Indeed it is a rather contrived example, though it requires to
a) perform the calculation of the cummulative sum (which means we have to
remember the previous value)
b) do a rather simple currnet value - previous (NON NULL) value
c) and note and update the current max of b)

To do all this we still use an array as a state, we do need to take into account
that the first value might not be null so we need to update it the first time,
i.e. the initial condition for the previous value needs to be NULL. If it were 0
and the first random value was negative the first step would be negative value
instead of 0.

Also update the awkward float8 to double precision.